python-backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.7
requests==2.31.0
Werkzeug==2.3.7
python-socketio==5.9.0
Flask-SocketIO==5.3.6
gunicorn==21.2.0
Flask-RESTful==0.3.10 